Re-worked Auto-Migrations of stored procs
Was having an issue with Auto-Migrations of stored procs acutally working (the type of the resource was written to CosmosDB instead of the content) once nuget installed into a project.  Changed from a project resource to an embedded resource.

Also some small clean-up work with deleting the database on Init if selected, as well as NotFound error handling.